adds discord log upload, avatar, username and log trimming
You can use `send-logfile-on` in the config decide when to the log file.

Copied the trimming funciton from `send_email`, if the log is to long (>1800 chars) and the log files is not sent.

You can use `username` and `avatar-url` in the config to customise the apperance of the message of the webhook.